backend/requirements.txt
fastapi
uvicorn
motor
python-multipart
pyjwt
qrcode[pil]
requests
cryptography
python-dotenv